requirements.txt
alembic==1.10.4
astroid==2.15.5
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dill==0.3.6
filelock==3.12.0
flake8==6.0.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
future==0.18.3
gevent==22.10.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
parso==0.8.3
pep8==1.7.1
platformdirs==3.5.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pylint==2.17.4
PyMySQL==1.0.3
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.30.0
requests-file==1.5.1
six==1.16.0
SQLAlchemy==2.0.13
tldextract==3.4.1
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.4
wrapt==1.15.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
